Yields and angular correlation loop implemented
Code calculates a 3D vector vYields that contains the yields obtained from subtracting the normalised OS and SS correlations, as defined in the configurations.json file. The structure is implemented as vYields[TUNE][ASSOCIATE][DEPENDENCY].
The user also has the option to show the angular spectra of the OS, SS, and correlated contributions.