First balancing plot produced
Legend still needs to be fixed and error propagation still needs to be done... but the first result is there. Shouldn't be too difficult to make the rest.
The code is anyways not the most 'efficient' with the many loops over the same datasets, but it will be very straightforward and clear to understand (no back and forward referencing to yield values; all of them are known from the beginning of the plotting procedure..). Also easy to make changes and add/remove stuff for the user without delving into the code (because they have the configuration.json)