Gave up on errors for now
It seems like the best (only?) way is to go from the production output.. It's a bit sad because it would be more challenging to provide these files on root. Anyways, I'll think about it.
Implementing the errors is not per-se difficult and can be done completely independently from the rest of the code.. should be OK in a later stage.
The subsampling from before on the outputs from PYTHIA will work (on simulation, batched level). However, as said above, it will be much less flexible this way, requiring a new full simulation even for small changes. One can imagine it should be possible to randomly sample from the existing histograms?