Added error calculation for balancing
Error calculation for the ratio (or any specific plots) needs to be done still. This will not be able to be generalised, because the user can come up with as creative (and extreme) as they want for their plot. It's not a problem, anyways, as for new plots one really needs to get into the code to add a plotting function. Error propagation still needs to be done for the baryon/meson ratio plots.